// src/types/models/CreateRoomResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateRoomResponse
 */
export interface CreateRoomResponse {
    /**
     * 
     * @type {string}
     * @memberof CreateRoomResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof CreateRoomResponse
     */
    roomKey: string;
}

/**
 * Check if a given object implements the CreateRoomResponse interface.
 */
export function instanceOfCreateRoomResponse(value: object): value is CreateRoomResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('roomKey' in value) || value['roomKey'] === undefined) return false;
    return true;
}

export function CreateRoomResponseFromJSON(json: any): CreateRoomResponse {
    return CreateRoomResponseFromJSONTyped(json, false);
}

export function CreateRoomResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateRoomResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'roomKey': json['roomKey'],
    };
}

  export function CreateRoomResponseToJSON(json: any): CreateRoomResponse {
      return CreateRoomResponseToJSONTyped(json, false);
  }

  export function CreateRoomResponseToJSONTyped(value?: CreateRoomResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'roomKey': value['roomKey'],
    };
}