// src/types/models/JoinRoomRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface JoinRoomRequest
 */
export interface JoinRoomRequest {
    /**
     * 
     * @type {string}
     * @memberof JoinRoomRequest
     */
    roomId: string;
    /**
     * 
     * @type {boolean}
     * @memberof JoinRoomRequest
     */
    isViewer?: boolean;
}

/**
 * Check if a given object implements the JoinRoomRequest interface.
 */
export function instanceOfJoinRoomRequest(value: object): value is JoinRoomRequest {
    if (!('roomId' in value) || value['roomId'] === undefined) return false;
    return true;
}

export function JoinRoomRequestFromJSON(json: any): JoinRoomRequest {
    return JoinRoomRequestFromJSONTyped(json, false);
}

export function JoinRoomRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): JoinRoomRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'roomId': json['roomId'],
        'isViewer': json['isViewer'] == null ? undefined : json['isViewer'],
    };
}

  export function JoinRoomRequestToJSON(json: any): JoinRoomRequest {
      return JoinRoomRequestToJSONTyped(json, false);
  }

  export function JoinRoomRequestToJSONTyped(value?: JoinRoomRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roomId': value['roomId'],
        'isViewer': value['isViewer'],
    };
}