// src/types/models/JoinRoomResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bodoge API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Chat } from './Chat';
import {
    ChatFromJSON,
    ChatFromJSONTyped,
    ChatToJSON,
    ChatToJSONTyped,
} from './Chat';

/**
 * 
 * @export
 * @interface JoinRoomResponse
 */
export interface JoinRoomResponse {
    /**
     * 
     * @type {Array<Chat>}
     * @memberof JoinRoomResponse
     */
    chatHistory: Array<Chat>;
    /**
     * 
     * @type {string}
     * @memberof JoinRoomResponse
     */
    roomKey: string;
}

/**
 * Check if a given object implements the JoinRoomResponse interface.
 */
export function instanceOfJoinRoomResponse(value: object): value is JoinRoomResponse {
    if (!('chatHistory' in value) || value['chatHistory'] === undefined) return false;
    if (!('roomKey' in value) || value['roomKey'] === undefined) return false;
    return true;
}

export function JoinRoomResponseFromJSON(json: any): JoinRoomResponse {
    return JoinRoomResponseFromJSONTyped(json, false);
}

export function JoinRoomResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): JoinRoomResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'chatHistory': ((json['chatHistory'] as Array<any>).map(ChatFromJSON)),
        'roomKey': json['roomKey'],
    };
}

  export function JoinRoomResponseToJSON(json: any): JoinRoomResponse {
      return JoinRoomResponseToJSONTyped(json, false);
  }

  export function JoinRoomResponseToJSONTyped(value?: JoinRoomResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'chatHistory': ((value['chatHistory'] as Array<any>).map(ChatToJSON)),
        'roomKey': value['roomKey'],
    };
}